docker/files/opt/mycroft-mark2-hardware/requirements.txt
adafruit-circuitpython-neopixel
dbus-next
